introduce datasources as centralized locations for refreshing, add announcements

+++ HackIllinois/Formatters.swift
@@ -44,49 +44,48 @@ extension Date {
         switch timeSinceDate {
         // YEARS
         case (2 * YEAR_IN_SECONDS)... :
-            let years = floor(timeSinceDate/YEAR_IN_SECONDS)
+            let years = Int(floor(timeSinceDate/YEAR_IN_SECONDS))
             return "\(years) years ago"
         case (1 * YEAR_IN_SECONDS)... :
             return "last year"
 
         // MONTH
         case (2 * MONTH_IN_SECONDS)... :
-            let month = floor(timeSinceDate/MONTH_IN_SECONDS)
+            let month = Int(floor(timeSinceDate/MONTH_IN_SECONDS))
             return "\(month) months ago"
         case (1 * MONTH_IN_SECONDS)... :
             return "last month"
 
         // WEEK
         case (2 * WEEK_IN_SECONDS)... :
-            let week = floor(timeSinceDate/WEEK_IN_SECONDS)
+            let week = Int(floor(timeSinceDate/WEEK_IN_SECONDS))
             return "\(week) weeks ago"
         case (1 * WEEK_IN_SECONDS)... :
             return "last week"
 
         // DAY
         case (2 * DAY_IN_SECONDS)... :
-            let day = floor(timeSinceDate/DAY_IN_SECONDS)
+            let day = Int(floor(timeSinceDate/DAY_IN_SECONDS))
             return "\(day) days ago"
         case (1 * DAY_IN_SECONDS)... :
             return "yesterday"
 
         // HOUR
         case (2 * HOUR_IN_SECONDS)... :
-            let hour = floor(timeSinceDate/HOUR_IN_SECONDS)
+            let hour = Int(floor(timeSinceDate/HOUR_IN_SECONDS))
             return "\(hour) hours ago"
         case (1 * HOUR_IN_SECONDS)... :
             return "yesterday"
 
         // MINUTE
         case (2 * MINUTE_IN_SECONDS)... :
-            let minute = floor(timeSinceDate/MINUTE_IN_SECONDS)
+            let minute = Int(floor(timeSinceDate/MINUTE_IN_SECONDS))
             return "\(minute) minutes ago"
         case (1 * MINUTE_IN_SECONDS)... :
             return "a minute ago"
 
         // SECOND
         case (15 * SECOND_IN_SECONDS)... :
-//            let second = floor(timeSinceDate/SECOND_IN_SECONDS)
             return "a few seconds ago"
 
         default:
